refactor : MockURLString 양식에 맞춰 재생성 (#345)
