--- KCS/KCSUnitTest/Mock/Entity/MockURLString.swift
@@ -1,17 +1,0 @@
-//
-//  MockURLString.swift
-//  KCSUnitTest
-//
-//  Created by 김영현 on 2/29/24.
-//
-
-import Foundation
-
-enum MockURLString: String {
-    
-    case cache = "test_cache_image_URL"
-    case noCache = "test_no_cache_image_URL"
-    case fail = "url"
-    case wrongURL = ""
-    
-}
